requirements.txt
absl-py==1.0.0
astunparse==1.6.3
autopep8==1.6.0
bcrypt==3.2.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==6.3.2
cryptography==37.0.1
fabric==2.7.0
filetype==1.0.13
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
<<<<<<< HEAD
=======
greenlet==1.1.2
>>>>>>> boilerplate
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
<<<<<<< HEAD
=======
invoke==1.7.0
>>>>>>> boilerplate
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.6
MarkupSafe==2.1.1
<<<<<<< HEAD
numpy
oauthlib==3.2.0
opt-einsum==3.3.0
=======
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
paramiko==2.10.4
pathlib2==2.3.7.post1
>>>>>>> boilerplate
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
<<<<<<< HEAD
=======
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.20.0
>>>>>>> boilerplate
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
<<<<<<< HEAD
=======
SQLAlchemy==1.4.36
>>>>>>> boilerplate
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.0
<<<<<<< HEAD
=======
WTForms==3.0.1
>>>>>>> boilerplate
zipp==3.8.0
